Updated Average Sentiment Crawl
Still in progress, I've basically gone for a greedy solution where we update the best sentiment found on the links from a parent link in the URL seed set, once done I call a single sentiment crawl on the optimal parent link based on the sentiment score of its overall links. Will refine further as looks to be a bug somewhere in the method, see the following for the current results from this implementation:

Crab\src\com\company\CrabResults.csv

Crab\src\com\company\AverageOptimal.csv

Also added back in the Serialization feature.