Updates to Data Structures
Made changes to the Linked List to use a Sentiment value/node instead of head and tail will allow to reduce the number of constant time factors given we are using somewhat large enough URL seed sets.

Stack also updated to reflect this and removed irrelevant methods that are not in use.