Update Utility.java
formatting is better for the serialization, however currently have a bug where only half of the JSON is being printed out and don't have an ending bracket. This traces back to why an error occurs if you call the deserialize_json function not sure if its the code or gson itself.