Added in new Stack implementation
a Lock-Free stack based on Treiber's design currently testing before integrating into the actual crawling method as the url stack.

the Lock-Free stack should reduce some of the contention seen with the safe push and pop of the original stack used since the beginning of the project, given the amount of threads trying to acquire the lock.