Resolved bugs in JSON serialization/deserialization methods
Previous method implementations had a bug were only so much of the data structure was being written to the JSON file. This method gets around that by converting the JSON format into a string and then writing/reading the string from the file into the data structure.

also added are methods to serialize/deserialize the visit list through gson.