Adjusted Thread pool attributes
Crab now calculates a core thread pool size automatically based off the number of cores available. the default is to take half of the cores but this can be changed.

Also updated is checking a URL off the stack to see whether its part of the blocked list this allows better filtering so we don't even consider letting Crab look at the given URL (i.e. facebook) and wasting cycles.

Note you can change the thread pool characteristics, I have found on my machine getting the right combination is somewhat difficult if this occurs feel free to switch to a fixed thread pool.