Added first part of new Keyword Crawl
Keyword Crawl has been redone, this version uses CoreNLP's POS tagger and searches for specific tags given by the user or goes by a standard default.

Adding a speech tag is done by:
Crab.tagsToSearch.add("NNP");

Crab searches through the title of the page and applies POS. Once the POS tags are found it then searches through a data structure (tag,word) looks up a tag for the keyword we are looking for and see's if it's in the values to the key (POS tag).

Note the download queue is for the second part where once keywords are found for an article it gets put into a queue for full sentiment analysis.

Results are currently written to CSV where you would see something like:
Site X, Keyword etc

only pages where we get a keyword match are documented.