Added Callable for Full Sentiment
This callable will take/"download" a web page using Jsoup, extract the body of text and then apply sentiment analysis then signify whether a result was obtained from the Sentiment Callable class.

The intended use is for the download queue within the main processing loop Crab will poll a URL that has keywords contained within it and then use this class to derive the overall sentiment and then record the result.