Added capability to execute batches of full sentiment tasks
For keyword crawl,  the method runFullSentimentTasks uses a completion service to process a queue of full sentiment tasks that are initiated through DonwloadCallable.

Crab only execute this when the queue is bigger than the variable execution_threshold this can be tuned to whatever or just on condition that the queue is not empty on each iteration of the main while loop.

Some amendments still to be made, I will add in a tagging capability so each full sentiment is tagged with the keywords that Crab has found in the last stage of the keyword crawl.