Added Full sentiment to run on Keyword Runnable at the moment
Added the capability to run full sentiment from the runnable class by setting the associated the bool value in the crab class.

Note this is only a quick fix and will most likely be deprecated once I add in tagging and have the right work scheduling in place.

also best to add JVM optimisation flags and aim to run for full capacity given the full sentiment can be intensive depending on the length of the overall article text extracted.