Added in feature to print all optimal results for a page
the standard default only prints the last considered "best" link for a page, ideally want more detailed so an equally positive sentiment for a variety of links looked at can get printed when running an optimal crawl.

This can now be done by setting the record all flag when in an optimal crawl.