Resolved blocked URL exceptions
Previously for the optimal crawl, a blocked url was connected through Jsoup before it was connected which depending on the site causes an exception to be throw. Instead the blocked URL is checked first before the jsoup method is initiated.

This resolves the HTTP status exceptions but can still occur depending on the link.

URI can throw a similar exception but I have noticed this is somewhat rare and dependent on the link that's been derived from the page. If this becomes more apparent on more crawls then I will implement a matching part when the link is derived rather than calling URI first.