Removed FullSentimentRunnable
Following some analysis this runnable has been removed, I am planning to implement this again but perhaps using Java's fork and join capability. Given their is sufficient lag on performance as full sentiment tasks build up given the size of some of the extracted text. In the meantime other features will be made use of in the keyword crawl.