Added results for keyword crawl
For this crawl, I added keywords based on four cryptocurrencies.

Visit list shows all pages looked at by Crab to find matches.

kword_matches is all the matches for the keywords that were inputted.

sentiment results is the sentiment analysis of the headline of the page that matches the keyword.