updated to handle HTTP status exceptions rather than throw them
Rather throwing the exception when crab encounters a http error code instead it now checks the status code it gets from Jsoup before continuing allowing to actually handle the exception rather than throwing it.

The status code array is filled with error codes that Crab has encountered since the project started. OptimalRunnable will be updated with this feature to.

Although line 209 could just be used it will eventually encounter the exception further down if not checked.