Keyword Crawl now does BFS
the new frontier technique utilises a priority queue where links with a higher amount of keyword matches go to the top of the queue for a best-first search like crawl.

From trial runs this gives a more detailed crawl for the keywords provided, sentiment runnable will also see such changes.

To utilise this method you have to set the flag FULL_DEPTH to true, otherwise you'll get a standard keyword crawl. the difference being the standard won't be as detailed and won't explore links further where as FULL_DEPTH will.

CrabTag is used to quantify each URL with its keyword or sentiment level to allow a comparator to be used for the priority queue.