Sentiment Crawl now updated, some minor updates to syntax
Sentiment Crawl now uses a frontier that prioritises crawling links with a higher sentiment score in their anchor text. Note that with the sentiment crawl if you do not set a crawl limit it will crawl until it is exhausted.

content extraction will be the next part of the sentiment crawl once database functionality is added.